Write test for add function
