+++ app/modules/checkboxes.js
@@ -1,43 +1,34 @@
-import { todoList, checkboxes } from './selectors.js';
+import { checkboxes } from './selectors.js';
 
-let tasks;
-function checkState() {
-  todoList.addEventListener('click', (e) => {
-    if (
-      localStorage.getItem('tasks') !== null
-      && localStorage.getItem('tasks') !== '[]'
-    ) {
-      tasks = JSON.parse(localStorage.getItem('tasks'));
-    }
-    if (e.target.nodeName === 'INPUT') {
-      const element = e.target;
-      const item = element.parentElement.innerText;
+// change checkbox state when checked or unchecked
+export default function checkState(element) {
+  const item = element.parentElement.children[1].innerText;
+  const tasks = localStorage.getItem('tasks')
+    ? JSON.parse(localStorage.getItem('tasks'))
+    : [];
+  if (element.checked) {
+    element.parentElement.classList.add('completed');
 
-      if (element.checked) {
-        element.parentElement.classList.add('completed');
-
-        const editedArr = tasks.map((task) => {
-          if (task.desc === item) {
-            //  return modified object property
-            return { ...task, completed: element.checked };
-          }
-          return task;
-        });
-        localStorage.setItem('tasks', JSON.stringify(editedArr));
-      } else {
-        element.parentElement.classList.remove('completed');
-
-        const editedArr = tasks.map((task) => {
-          if (task.desc === item) {
-            //  return modified object property
-            return { ...task, completed: false };
-          }
-          return task;
-        });
-        localStorage.setItem('tasks', JSON.stringify(editedArr));
+    const editedArr = tasks.map((task) => {
+      if (task.desc === item) {
+        //  return modified object property
+        return { ...task, completed: element.checked };
       }
-    }
-  });
+      return task;
+    });
+    localStorage.setItem('tasks', JSON.stringify(editedArr));
+  } else {
+    element.parentElement.classList.remove('completed');
+
+    const editedArr = tasks.map((task) => {
+      if (task.desc === item) {
+        //  return modified object property
+        return { ...task, completed: false };
+      }
+      return task;
+    });
+    localStorage.setItem('tasks', JSON.stringify(editedArr));
+  }
 }
 
 // check if checkbox is checked on load then style the checked task
@@ -50,4 +41,4 @@ function checkOnLoad() {
   });
 }
 
-export { checkState, checkOnLoad };
+export { checkOnLoad };
